create custom field for services
